feat: Made dtype dynamic in models

+++ doctr/models/recognition/sar/pytorch.py
@@ -107,7 +107,7 @@ class SARDecoder(nn.Module):
         # initialize states (each of shape (N, rnn_units))
         hx = [None, None]
         # Initialize with the index of virtual START symbol (placed after <eos>)
-        symbol = torch.zeros((features.shape[0], self.vocab_size + 1), device=features.device)
+        symbol = torch.zeros((features.shape[0], self.vocab_size + 1), device=features.device, dtype=features.dtype)
         logits_list = []
         for t in range(self.max_length + 1):  # keep 1 step for <eos>
 
@@ -131,7 +131,7 @@ class SARDecoder(nn.Module):
                 _symbol = gt[:, t]  # type: ignore[index]
             else:
                 _symbol = logits.argmax(-1)
-            symbol = F.one_hot(_symbol, self.vocab_size + 1).to(dtype=torch.float32)
+            symbol = F.one_hot(_symbol, self.vocab_size + 1).to(dtype=features.dtype)
             logits_list.append(logits)
         outputs = torch.stack(logits_list, 1)  # shape (N, max_length + 1, vocab_size + 1)
